backend/pyproject.toml
# Copyright (c) 2026 Brendan Bank
# SPDX-License-Identifier: BSD-2-Clause

[project]
name = "atrium-backend"
version = "0.1.0"
description = "Atrium — auth/RBAC/notifications/audit starter (backend)"
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0.36",
    "aiomysql>=0.2.0",
    "alembic>=1.14",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "python-dotenv>=1.0",
    "structlog>=24.4",
    "email-validator>=2.2",
    "httpx>=0.27",
    "jinja2>=3.1",
    "aiosmtplib>=3.0",
    "apscheduler>=3.10",
    "fastapi-users[sqlalchemy]>=13.0",
    "bcrypt>=4.2",
    "pyjwt>=2.10",
    "python-multipart>=0.0.12",
    "bleach>=6.3",
    "pyotp>=2.9",
    "webauthn>=2.4",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "testcontainers[mysql]>=4.8",
    "ruff>=0.7",
    "mypy>=1.13",
    "types-python-dateutil",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.flake8-bugbear]
# FastAPI's dependency-injection pattern relies on Depends(...) in
# default values. Query/Path/Body/etc. are the same story. Exempt them
# from B008 "no function call in defaults" so idiomatic FastAPI routes
# aren't flagged.
extend-immutable-calls = [
    "fastapi.Depends",
    "fastapi.Query",
    "fastapi.Path",
    "fastapi.Body",
    "fastapi.Header",
    "fastapi.Cookie",
    "fastapi.Form",
    "fastapi.File",
    "fastapi.Security",
    # Our own dependency factory; same rule as Depends — safe to
    # invoke at import time in default args.
    "app.auth.rbac.require_perm",
]

[tool.ruff.lint.per-file-ignores]
# Migrations are boilerplate templates — keep them readable over strict
# rules about type aliases and unused args.
"alembic/versions/*.py" = ["UP007", "UP035"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers"
markers = [
    "real_2fa: exercise the real 2FA gate; conftest's _auto_pass_2fa bypass is skipped",
]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
